double -> int

+++ src/main/java/edu/hw2/task2/Rectangle.java
@@ -25,7 +25,7 @@ public class Rectangle {
         return new Rectangle(width, height);
     }
 
-    double area() {
+    int area() {
         return width * height;
     }
 }
